tidy(runtime-api): extract request dispatch from socket transport [WIP]
Moves the request-domain layer out of src/runtime-api/server.py into
src/runtime-api/dispatcher.py. server.py keeps daemon config, Unix-socket
ownership/permissions, NDJSON framing, connection lifecycle and supervision.

  src/runtime-api/server.py       499 -> 146
  src/runtime-api/dispatcher.py     0 -> 416  (new)

The move is verbatim: every original line is present in one file or the other
except the three call sites intentionally rewired to self.dispatcher.*
(handle / recover / resolver_loop). Verified by line-set comparison against
main, not by eye.

Actor identity stays daemon-resolved in RuntimeServer.__init__ and is passed to
the dispatcher explicitly, so a client parameter still cannot override it.

Two things this surfaced, both fixed:

1. `sqlite3` was used but not imported in the extracted module. py_compile and
   four of five suites passed anyway — the reference is on the idempotency-race
   path (`except sqlite3.IntegrityError`), so it only fires on a SECOND
   execution against the same approval. The e2e suite caught it as
   "an approval authorizes exactly ONE execution (consumed)" failing
   deterministically (main 2/2 pass, extraction 2/2 fail). A security contract
   held only by an exception path, invisible to compile and unit tests.

2. The executor lookup read the module-global EXECUTORS while the constructor
   accepted an `executors` argument — so injecting a map would have been
   silently ignored. Now reads self.executors, making the seam real.

Existing suites retargeted to drive the dispatcher (45 call sites); they were
testing domain behavior that now lives there. All five pass:
runtime-api-server, -store, -issue-failure, -rundir, -e2e.

WIP — NOT ready for review. Still outstanding per the handoff:
- tests/runtime-api-dispatcher.test.py (~20 required direct cases)
- CLAUDE.md rule, AGENTS.md, architecture-boundaries.md, src-map.md
- the disposable-socket live-path verification
